Refactor and add remapping tests
